pyproject.toml
[project]
name = "telegram-leetcode-bot"
version = "0.1.0"
description = "A Telegram bot that delivers daily LeetCode questions and provides algorithm learning resources"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "dotenv>=0.9.9",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "apscheduler>=3.10.0",
]
